+++ packages/winnow/lib/filter-and-transform/transforms/project.js
@@ -1,21 +1,22 @@
 const projectCoordinates = require('../../helpers/project-coordinates')
 
-function project (geometry, targetCoordinateSystem) {
+function project (geometry, sourceCoordinateSystem, targetCoordinateSystem) {
   if (!geometry || !targetCoordinateSystem) return geometry
 
   const { type, coordinates: sourceCoordinates } = geometry
 
   if (!type || !sourceCoordinates) return geometry
 
-  return tryProjectingGeometry({ type, sourceCoordinates, targetCoordinateSystem })
+  return tryProjectingGeometry({ type, sourceCoordinates, sourceCoordinateSystem, targetCoordinateSystem })
 }
 
-function tryProjectingGeometry ({ type, sourceCoordinates, targetCoordinateSystem }) {
+function tryProjectingGeometry ({ type, sourceCoordinates, sourceCoordinateSystem, targetCoordinateSystem }) {
   try {
     return {
       type,
       coordinates: projectCoordinates({
         coordinates: sourceCoordinates,
+        fromSR: sourceCoordinateSystem,
         toSR: targetCoordinateSystem
       })
     }