+++ packages/winnow/lib/query/package-features.js
@@ -1,6 +1,7 @@
 const _ = require('lodash')
+const normalizeSpatialReference = require('../helpers/normalize-spatial-reference')
 
-function packageFeatures (features = [], { groupBy, aggregates, collection } = {}) {
+function packageFeatures (features = [], { groupBy, aggregates, collection, projection } = {}) {
   if (groupBy || (!aggregates && !collection)) {
     return features
   }
@@ -10,7 +11,19 @@ function packageFeatures (features = [], { groupBy, aggregates, collection } = {
   }
 
   if (collection) {
-    return _.chain(collection).cloneDeep().set('features', features).value()
+    return packageCollection({ features, collection, projection })
   }
 }
+
+function packageCollection ({ features, collection, projection }) {
+  const outputCollection = _.chain(collection).cloneDeep().set('features', features)
+  if (!projection) {
+    return outputCollection.value()
+  }
+
+  const { wkid } = normalizeSpatialReference(projection)
+  const crs = `urn:ogc:def:crs:EPSG::${wkid}`
+  return outputCollection.set('crs.properties.name', crs).value()
+}
+
 module.exports = packageFeatures