+++ packages/winnow/test/unit/sql-query-builder/select/geometry-select-fragment.spec.js
@@ -15,7 +15,7 @@ test('createGeometryFragment: empty options', t => {
 
 test('createGeometryFragment: projection option', t => {
   const geometryFragment = createGeometrySelectFragment({ projection: 3857 })
-  t.equal(geometryFragment, 'project(geometry,?) as geometry')
+  t.equal(geometryFragment, 'project(geometry,?,?) as geometry')
   t.end()
 })
 
@@ -33,18 +33,18 @@ test('createGeometryFragment: toEsri option', t => {
 
 test('createGeometryFragment: projection and geometryPrecision options', t => {
   const geometryFragment = createGeometrySelectFragment({ projection: 3857, geometryPrecision: 1 })
-  t.equal(geometryFragment, 'reducePrecision(project(geometry,?),?) as geometry')
+  t.equal(geometryFragment, 'reducePrecision(project(geometry,?,?),?) as geometry')
   t.end()
 })
 
 test('createGeometryFragment: projection, geometryPrecision, and toEsri options', t => {
   const geometryFragment = createGeometrySelectFragment({ projection: 3857, geometryPrecision: 1, toEsri: true })
-  t.equal(geometryFragment, 'esriGeometry(reducePrecision(project(geometry,?),?)) as geometry')
+  t.equal(geometryFragment, 'esriGeometry(reducePrecision(project(geometry,?,?),?)) as geometry')
   t.end()
 })
 
 test('createGeometryFragment: projection, and toEsri options', t => {
   const geometryFragment = createGeometrySelectFragment({ projection: 3857, toEsri: true })
-  t.equal(geometryFragment, 'esriGeometry(project(geometry,?)) as geometry')
+  t.equal(geometryFragment, 'esriGeometry(project(geometry,?,?)) as geometry')
   t.end()
 })